requirements.txt
functions-framework==3.*
pandas
bs4
lxml
nltk
requests
google-cloud-storage